More consistent unit handling
There was an issue where `effective_wavelength` did not work on `PhotometricFilter` instances constructed without units on the wavelength vector. It essentially required some unit information to make sure the Vega spectrum wavelengths and the filter wavelengths were both the same. This caused additional problems because `effective_wavelength` is called in the `Base.show` implementation. In this commit, we still allow constructing `PhotometricFilter` instances without units on the input wavelengths, but we assume that they are Angstroms in this case (or, more specifically, whatever unit the `wave_unit` constant is ste to). We convert wavelengths to angstroms before writing them into the `PhotometricFilter` struct, so there is no unit information in the struct at all. Units are added in the accessor methods like `wave`, which are then used in the other methods as well. This also fixes a problem where the input dimensions of the wavelengths were not checked and so could be inconsistent (e.g., not `Unitful.Length`). Still need to do the `apply` methods ...